Basic Interperetation
Added equals and plus operations, used them to create a basic script which moves the cube to the right.

Shows functionality of the interperator as the operations are linked using blueprints, but execute dynamically during runtime.

Need further functionality to allow user controls (e.g. an evaluatable that detects a specific key is down and an executable (if statement) to check the value and react accordingly.