Made NumVar more generic
Added a system that allows variables to be more generic. When being accessed for the first time, each variable sets its references to block manager and var manager (done this lae to ensure all appropriate objects are created first).

After setting these links, the variable then calls a function in var manager which takes in the components name (set in the editor) and assigns it an enumerated value. This enum is then used when calling a generic SetNumber or GetNumber function in var manager. The use of enums allows for the appropriate checks to be done with a switch statement, which is faster than equating ifs, espescially when the list of variables gets longer.

A method involving function pointers was tried but proved to complicated for the limited timespan.