Connection Manager For Buttons
Expanded the connection manager to take in buttons so that they can be passed around to form visual connections using the line drawer.

Also implemented a basic validity check in If Statement to ensure it has a valid input node. This same system should be expanded to cover all types of inputs as it prevents the reading of garbage data if an object is destroyed and the input is not reset.