Connection Breaking
Connections are now broken when clicking on a slot.

Previously the visual connection would be broken but the pointer within the C++ object would still point to the last object it was connected to until a new connection was made. Now the pointers are set to null whenever they are given to the connection manager.