Created Vector Type
Created a new data type of Vector

Uses class "Evaluatable"  as base for "Vector2", which works similarly to the boolean and number data types already implemented

Vector2.h has a definition for a "MyVector" struct which is used as the underlying data type (the same way as Number uses float)

Also created a class to construct a Vector2 from 2 Number inputs and atached it to a corresponding UI widget.

Next step is to make the Shoot block take in a Vector2 as an input.