Started Implementing Play State
Started adding a system which allows for 2 distinct states: Scripting and Playing.

Previously, these both happened at the same time, but now you can only be doing 1 or the other.

This change allows for scripts to be set up without being spammed with messages about missing connections, as this is only checked in the playing stage and will immediately revert to the scripting stage if an error occurs.